Move Reference to separate window
